Remove superfluous synchronization
Turns out, the real problem all along wasn't that I was letting the
MainWindow::mainLoop run while I was creating the scene graph. The real problem
is that OSPRay rendering happens in an extra thread, which means that it will
see your unfinished data while you're creating it.

The fix is to simply disable rendering while creating the scene graph, and then
re-enable at the end.